1)Изменился стиль и положение кнопок 2)Layouts renamed 3)Initial values for SetPlayers added 4)Добавлено отображение частичек в таблице 5)Rubber class separated from Table
