+++ app/src/main/java/com/example/rubberbridge/SetPlayers.kt
@@ -1,14 +1,12 @@
 package com.example.rubberbridge
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import com.example.rubberbridge.databinding.FragmentFirstBinding
+import com.example.rubberbridge.databinding.FragmentSetPlayersBinding
 import java.io.File
 
 
@@ -17,7 +15,7 @@ import java.io.File
  */
 class SetPlayers : Fragment() {
 
-    private var _binding: FragmentFirstBinding? = null
+    private var _binding: FragmentSetPlayersBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -28,13 +26,16 @@ class SetPlayers : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
 
-        _binding = FragmentFirstBinding.inflate(inflater, container, false)
+        _binding = FragmentSetPlayersBinding.inflate(inflater, container, false)
         return binding.root
 
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+
+
 
         binding.buttonFirst.setOnClickListener {
             //var text_length = binding.editTextPlayer1.length()
