+++ app/src/main/res/layout/fragment_result_of_deal.xml
@@ -2,9 +2,9 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-tools:context=".ResultOfDeal">
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ResultOfDeal">
 
     <TextView
         android:id="@+id/textView5"
@@ -12,7 +12,7 @@ tools:context=".ResultOfDeal">
         android:layout_height="47dp"
         android:layout_marginStart="10dp"
         android:layout_marginTop="30dp"
-        android:text="номер пары:"
+        android:text="Команда:"
         android:textAppearance="@style/TextAppearance.AppCompat.Large"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView4" />
@@ -26,15 +26,25 @@ tools:context=".ResultOfDeal">
 
     <Button
         android:id="@+id/button_approve_contract"
-        android:layout_width="125dp"
-        android:layout_height="53dp"
-        android:layout_marginStart="24dp"
-        android:layout_marginTop="10dp"
-        android:text="контракт введён"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
 
-    <EditText
+        android:layout_marginTop="10dp"
+
+        android:background="@drawable/btn_style"
+
+        android:text="Записать игру"
+        app:backgroundTint="@color/blue"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.5"
+
+
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.9"/>
+
+   <EditText
         android:id="@+id/edit_suit"
         android:layout_width="80dp"
         android:layout_height="47dp"
@@ -43,7 +53,7 @@ tools:context=".ResultOfDeal">
         android:ems="10"
         android:inputType="number"
         app:layout_constraintStart_toEndOf="@+id/text_suit"
-        app:layout_constraintTop_toBottomOf="@+id/button_approve_contract" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <EditText
         android:id="@+id/edit_level"
@@ -82,22 +92,22 @@ tools:context=".ResultOfDeal">
         android:id="@+id/text_suit"
         android:layout_width="140dp"
         android:layout_height="47dp"
-        android:layout_marginStart="10dp"
         android:layout_marginTop="40dp"
+        android:layout_marginStart="8dp"
         android:layout_marginEnd="23dp"
-        android:text="масть:"
+        android:text="Масть:"
         android:textAppearance="@style/TextAppearance.AppCompat.Large"
         app:layout_constraintEnd_toStartOf="@+id/edit_suit"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/button_approve_contract" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/text_level"
         android:layout_width="140dp"
         android:layout_height="47dp"
-        android:layout_marginStart="10dp"
-        android:layout_marginTop="30dp"
-        android:text="уровень:"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="28dp"
+        android:text="Уровень:"
         android:textAppearance="@style/TextAppearance.AppCompat.Large"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/text_suit" />
@@ -106,17 +116,17 @@ tools:context=".ResultOfDeal">
         android:id="@+id/textView4"
         android:layout_width="140dp"
         android:layout_height="47dp"
-        android:layout_marginStart="10dp"
+        android:layout_marginStart="8dp"
         android:layout_marginTop="30dp"
-        android:text="взяток:"
+        android:text="Итог сдачи:"
         android:textAppearance="@style/TextAppearance.AppCompat.Large"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/text_level" />
 
     <TextView
         android:id="@+id/error_text"
-        android:layout_width="386dp"
-        android:layout_height="205dp"
+        android:layout_width="391dp"
+        android:layout_height="41dp"
         android:layout_marginStart="8dp"
         android:layout_marginTop="58dp"
         android:textSize="20sp"
