+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,8 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+
+
+
+    <color name="blue">#ff1565c0</color>
 </resources>
